feat: integrate email verification modal into login and registration flows

+++ src/lib/apiClient.ts
@@ -98,35 +98,39 @@ const processErrorResponse = (error: unknown): string => {
 
 // Function to Store Token After Login
 export const storeUserData = async (data: DataFromLogin) => {
-  try {
-    // Store in HTTP-only cookie (optional, requires backend endpoint)
-    sessionStorage.setItem("token", data.token);
-    const user: User = await getFn(`api/users/view/${data.user.id}`);
-    // const user: User = await getFn("api/account/currentuserdata");
-    if (user && user.email_verified_at) {
-      const response = await fetch("/api/auth/set-user", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ ...data, user }),
-        // body: JSON.stringify({ token: data.token }),
-      });
+  if (data.token) {
+    try {
+      // Store in HTTP-only cookie (optional, requires backend endpoint)
+      sessionStorage.setItem("token", data.token);
+      const user: User = await getFn(`api/users/view/${data.user.id}`);
+      // const user: User = await getFn("api/account/currentuserdata");
+      if (user && user.email_verified_at) {
+        const response = await fetch("/api/auth/set-user", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ ...data, user }),
+          // body: JSON.stringify({ token: data.token }),
+        });
 
-      if (!response.ok) {
-        throw new Error("Failed to set token cookie");
+        if (!response.ok) {
+          throw new Error("Failed to set token cookie");
+        }
+        getUser();
+        return user
+      } else if (user && !user.email_verified_at) {
+        sessionStorage.clear();
+        return null;
+      } else {
+        sessionStorage.clear();
+        return undefined;
       }
-      getUser();
-      return user
-    } else if (user && !user.email_verified_at) {
-      sessionStorage.clear();
-      return null;
-    } else {
-      sessionStorage.clear();
-      return undefined;
+    } catch (error) {
+      console.error("Failed to store token:", error);
     }
-  } catch (error) {
-    console.error("Failed to store token:", error);
+  } else {
+    return null;
   }
 };
